aa
aa

+++ DeadMoon-main/Assets/Scripts/CharacterStats.cs
@@ -5,10 +5,10 @@ using TMPro;
 
 public class CharacterStats : MonoBehaviour
 {
-    public int health;
+    public float health;
 
 
-   public int maxHealth;
+   public float maxHealth;
 
     [SerializeField] protected bool isDead;
 
@@ -41,23 +41,23 @@ public class CharacterStats : MonoBehaviour
         Destroy(gameObject);
     }
 
-    public void SetHealthTo(int healthSetTo)
+    public void SetHealthTo(float healthSetTo)
     {
         health = healthSetTo;
         CheckHealth();
 
     }
 
-    public virtual void TakeDamage(int damage)
+    public virtual void TakeDamage(float damage)
     {
-        int healtAfterDamage = health - damage;
+        float healtAfterDamage = health - damage;
         SetHealthTo(healtAfterDamage);
         
     }
 
-    public void Heal(int heal)
+    public void Heal(float heal)
     {
-        int healthAfterHeal = health + heal;
+        float healthAfterHeal = health + heal;
         SetHealthTo(healthAfterHeal);
     }
